refactor: 식당 생성 시 지역 상세 정보 추가를 위한 constituency_name 추가
