+++ src/main/java/com/mju/lighthouseai/domain/restaurant/dto/controller/RestaurantCreateControllerRequestDto.java
@@ -6,7 +6,9 @@ public record RestaurantCreateControllerRequestDto(
     Integer price,
     String menu,
     String opentime,
-    String closetime
+    String closetime,
+    String constituency_name
+
 ) {
 
 }
